Added optional parameter to plot.get_auto_camera method
Suppose the user has a very high aspect ratio mesh. In these situations the default visualization
might not be optimal. Further, the user might not be able to zoom in enough to visualize the
details of the mesh.

One approach to solve this problem is by setting the camera vector. The current version of K3D
exposes the get_auto_camera method which is useful to compute the camera vector. However, it
only considers the bounds of the objects added to the plot.

This commit adds the `bounds` keyword argument to the get_auto_camera method. If this parameter
is provided in the function call, it overrides the bounds of the objects in the plots.
Therefore, the user can easily call this method in order to set the camera position only by
specifying the appropriate bounds.