Added capability to set labels color
Previously, the label color was hard coded, with value `0x444444`. A keyword argument has been created,
`label_color=0x444444`, both in Python as well as Javascript.

There was a bug in `js/src/providers/threejs/objects/Text.js` in the `colorToHex` function, which
returned the string `"#aN" (supposedly meaning NaN, not a number). The problem has been resolved
by using the `getHexString` method of THREE.Color.