Fix coordinates order for implicit isosurface
Issue #340 showed that K3D-Jupyter isosurfaces are incorrectly visualized on the screen,
in comparison to Plotly and Sage math.

This commit swaps the order of the `x` and `z` coordinates, thanks to which the visualization
appears correct.

This edit has been tested with the examples found at:
https://doc.sagemath.org/html/en/reference/plot3d/sage/plot/plot3d/implicit_plot3d.html

K3D-Jupyter now plots isosurfaces with the correct orientation: the comparison was done
against both Plotly and Sage math.